requirements.txt
streamlit
pymongo
qdrant-client
pydantic
python-dotenv
bcrypt
argon2-cffi
sentence-transformers
pypdf
python-docx
requests
huggingface-hub
tqdm
uvicorn
fastapi
pyjwt
openai